% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiapp_run.R
\name{emojiapp_run}
\alias{emojiapp_run}
\title{Emoji Data Explorer}
\usage{
emojiapp_run()
}
\description{
Shiny app para explorar emojis por país latinoamericano.
}
\details{
Start emojidataexplorer
}
\examples{
\dontrun{
library(emojidataexplorer)
emojiapp::emojiapp_run()
}
}
\keyword{emoji}
\keyword{shiny,}
